[DEV-4926] reorg fiscal_year helpers
