+++ usaspending_api/download/management/commands/populate_monthly_files.py
@@ -8,7 +8,7 @@ from django.conf import settings
 from django.core.management.base import BaseCommand
 from usaspending_api.awards.v2.lookups.lookups import procurement_type_mapping, assistance_type_mapping
 from usaspending_api.common.helpers.dict_helpers import order_nested_object
-from usaspending_api.common.helpers.generic_helper import generate_fiscal_year
+from usaspending_api.common.helpers.fiscal_year_helpers import generate_fiscal_year
 from usaspending_api.common.sqs.sqs_handler import get_sqs_queue
 from usaspending_api.download.filestreaming import download_generation
 from usaspending_api.download.helpers import multipart_upload, pull_modified_agencies_cgacs
