+++ usaspending_api/search/v2/views/spending_over_time.py
@@ -23,15 +23,17 @@ from usaspending_api.common.experimental_api_flags import (
     is_experimental_elasticsearch_api,
     mirror_request_to_elasticsearch,
 )
-from usaspending_api.common.helpers.orm_helpers import FiscalMonth, FiscalQuarter, FiscalYear
-from usaspending_api.common.helpers.generic_helper import (
+from usaspending_api.common.helpers.fiscal_year_helpers import (
     bolster_missing_time_periods,
     generate_fiscal_date_range,
     generate_fiscal_month,
     generate_fiscal_year,
+)
+from usaspending_api.common.helpers.generic_helper import (
     get_generic_filters_message,
     min_and_max_from_date_ranges,
 )
+from usaspending_api.common.helpers.orm_helpers import FiscalMonth, FiscalQuarter, FiscalYear
 from usaspending_api.common.query_with_filters import QueryWithFilters
 from usaspending_api.common.validator.award_filter import AWARD_FILTER
 from usaspending_api.common.validator.pagination import PAGINATION
